[fix] Fix register size and update internal dtc

+++ src/bus/device_tree.rs
@@ -4,7 +4,7 @@ mod dts;
 use super::mrom::Mrom;
 
 impl Mrom {
-    pub fn load_dtb(&mut self, dram_addr: u32) {
+    pub fn load_dtb(&mut self, dram_addr: u64) {
         let dts: String = dts::make_dts(dram_addr).replace("  ", "");
         let dtb: Vec<u8> = dtb::make_dtb(dts);
         self.mrom.extend(dtb);
